refactor: DRY guard composition + vendor-agnostic gather extensions
Overview #4: Extract ComposeAndWrap(Guard) in ConditionSourceBuilder to
de-duplicate identical composition logic from Build(), BuildVsSource(),
and ArrayContains(). 45 lines of structural duplication → 12-line helper.
All operator methods now delegate to a single composition point.

Overview #5: Move typed Include<TComponent, TModel> gather extensions
from Native/Fusion to core GatherExtensions with IComponent+IInputComponent
constraint. Eliminates FusionGatherExtensions entirely. Keep untyped
Include<TModel> (defaults to NativeTextBox) in NativeGatherExtensions.
Third vendor = zero gather work.
