+++ Alis.Reactive.Native/Extensions/NativeGatherExtensions.cs
@@ -7,8 +7,8 @@ using Alis.Reactive.Native.Components;
 namespace Alis.Reactive.Native.Extensions
 {
     /// <summary>
-    /// Adds Include() and Include&lt;TComponent&gt;() to GatherBuilder for native DOM components.
-    /// Vendor and readExpr derived from component instances — no hardcoded strings.
+    /// Native-specific gather shorthand. The untyped Include() defaults to NativeTextBox.
+    /// Typed Include&lt;TComponent&gt;() overloads live in core GatherExtensions (vendor-agnostic).
     /// </summary>
     public static class NativeGatherExtensions
     {
@@ -30,46 +30,6 @@ namespace Alis.Reactive.Native.Extensions
                 _defaultComponent.Vendor,
                 propertyName,
                 _defaultComponent.ReadExpr));
-            return self;
-        }
-
-        /// <summary>
-        /// Gathers the value of a native DOM component bound to a model property.
-        /// ReadExpr is resolved from the component type's IInputComponent.ReadExpr property.
-        /// </summary>
-        public static GatherBuilder<TModel> Include<TComponent, TModel>(
-            this GatherBuilder<TModel> self,
-            Expression<Func<TModel, object?>> expr)
-            where TComponent : NativeComponent, IInputComponent, new()
-            where TModel : class
-        {
-            var component = new TComponent();
-            var elementId = IdGenerator.For<TModel>(expr);
-            var propertyName = ExpressionPathHelper.ToPropertyName(expr);
-            self.AddItem(new ComponentGather(
-                elementId,
-                component.Vendor,
-                propertyName,
-                component.ReadExpr));
-            return self;
-        }
-
-        /// <summary>
-        /// Gathers the value of a native DOM component identified by string ref.
-        /// </summary>
-        public static GatherBuilder<TModel> Include<TComponent, TModel>(
-            this GatherBuilder<TModel> self,
-            string refId,
-            string name)
-            where TComponent : NativeComponent, IInputComponent, new()
-            where TModel : class
-        {
-            var component = new TComponent();
-            self.AddItem(new ComponentGather(
-                refId,
-                component.Vendor,
-                name,
-                component.ReadExpr));
             return self;
         }
     }
