+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/AllModulesTogether/Architecture/Index.cshtml
@@ -4,7 +4,6 @@
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
 @using Alis.Reactive.Fusion.Components
-@using Alis.Reactive.Fusion.Extensions
 @using Alis.Reactive.Descriptors.Sources
 @using Alis.Reactive.Descriptors
 @using Alis.Reactive.Descriptors.Commands
