+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/AllModulesTogether/ComponentGather/Index.cshtml
@@ -2,7 +2,6 @@
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
 @using Alis.Reactive.Fusion.Components
-@using Alis.Reactive.Fusion.Extensions
 @using Alis.Reactive.SandboxApp.Areas.Sandbox.Models
 @using Syncfusion.EJ2.DropDowns
 @using Microsoft.AspNetCore.Mvc.Rendering
