+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/AllModulesTogether/IdGenerator/Index.cshtml
@@ -3,7 +3,6 @@
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
 @using Alis.Reactive.Fusion.Components
-@using Alis.Reactive.Fusion.Extensions
 @{
     ViewData["Title"] = "IdGenerator Test";
