+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/AllModulesTogether/Todo/Index.cshtml
@@ -1,7 +1,6 @@
 @model TodoModel
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
-@using Alis.Reactive.Fusion.Extensions
 @using Alis.Reactive.Fusion.Components
 @using Alis.Reactive.Fusion.AppLevel
 @using Alis.Reactive.SandboxApp.Areas.Sandbox.Models
