+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/Components/Fusion/DropDownList/Index.cshtml
@@ -2,7 +2,6 @@
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
 @using Alis.Reactive.Fusion.Components
-@using Alis.Reactive.Fusion.Extensions
 @{
     ViewData["Title"] = "DropDownList";
