+++ Alis.Reactive.SandboxApp/Areas/Sandbox/Views/_ViewImports.cshtml
@@ -1,6 +1,7 @@
 @using Alis.Reactive.SandboxApp
 @using Alis.Reactive.SandboxApp.Models
 @using Alis.Reactive.SandboxApp.Areas.Sandbox.Models
+@using Alis.Reactive.Builders.Requests
 @using Syncfusion.EJ2
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Alis.Reactive.NativeTagHelpers
