+++ examples/resident-intake/Views/_ViewImports.cshtml
@@ -3,7 +3,7 @@
 @using Alis.Reactive.Native.Extensions
 @using Alis.Reactive.Native.Components
 @using Alis.Reactive.Native.AppLevel
-@using Alis.Reactive.Fusion.Extensions
+@using Alis.Reactive.Builders.Requests
 @using Alis.Reactive.Fusion.Components
 @using Alis.Reactive.Fusion.AppLevel
 @using Syncfusion.EJ2
